+++ source/core/Main.js
@@ -94,7 +94,6 @@ export {PerspectiveCamera} from "./objects/cameras/PerspectiveCamera.js";
 export {OrthographicCamera} from "./objects/cameras/OrthographicCamera.js";
 
 export {AudioEmitter} from "./objects/audio/AudioEmitter.js";
-export {PositionalAudio} from "./objects/audio/PositionalAudio.js";
 
 export {Script} from "./objects/script/Script.js";
 export {NodeScript} from "./objects/script/NodeScript.js";